miner/pearl-gateway/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "pytest-asyncio>=1.2.0",
  "pytest-xdist>=3.8.0",
  "pytest>=8.3.0",
  "ruff>=0.13.0"
]

[project]
name = "pearl-gateway"
version = "0.1.0"
description = "The Pearl Gateway miner for the Pearl project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.10.0",
  "bitcoin-utils>=0.7.0",
  "blake3>=1.0.7",
  "fastjsonschema>=2.16.0",
  "miner-utils",
  "numpy>=1.20.0",
  "py-pearl-mining",
  "prometheus-client>=0.23.1",
  "pybase64>=1.4.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "pyyaml>=6.0",
  "torch==2.9.1",
]

[project.scripts]
pearl-gateway = "pearl_gateway.cli:main"

